Start using namespaces and subdirectories for better code organization
