--- include/IPCClient.h
@@ -1,16 +1,0 @@
-//
-// Created by Antoine on 08/03/2025.
-//
-
-#ifndef IPCCLIENT_H
-#define IPCCLIENT_H
-
-
-
-class IPCClient {
-
-};
-
-
-
-#endif //IPCCLIENT_H
