+++ include/input/IPCClient.h
@@ -1,0 +1,10 @@
+#ifndef IPCCLIENT_H
+#define IPCCLIENT_H
+
+namespace input {
+    class IPCClient {
+
+    };
+}
+
+#endif //IPCCLIENT_H
