--- src/OverlayWindow.cpp
@@ -1,5 +1,0 @@
-//
-// Created by Antoine on 08/03/2025.
-//
-
-#include "OverlayWindow.h"
